requirements.txt
Django==2.0.5
django-filter==1.1.0
django-mssql==1.8
django-pyodbc-azure==2.0.4.1
djangorestframework==3.8.2
Markdown==2.6.11
pyodbc==4.0.23
pytz==2018.4